Moved seeding function to it's own file for separation of concerns, created seeding script
